Update to new HDF namespace

+++ uk.ac.gda.analysis/src/uk/ac/gda/analysis/hdf5/Hdf5Helper.java
@@ -29,11 +29,11 @@ import org.eclipse.dawnsci.hdf5.HDF5Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ncsa.hdf.hdf5lib.H5;
-import ncsa.hdf.hdf5lib.HDF5Constants;
-import ncsa.hdf.hdf5lib.exceptions.HDF5Exception;
-import ncsa.hdf.hdf5lib.structs.H5G_info_t;
-import ncsa.hdf.object.h5.H5Datatype;
+import hdf.hdf5lib.H5;
+import hdf.hdf5lib.HDF5Constants;
+import hdf.hdf5lib.exceptions.HDF5Exception;
+import hdf.hdf5lib.structs.H5G_info_t;
+import hdf.object.h5.H5Datatype;
 
 /**
  *
@@ -298,7 +298,7 @@ public class Hdf5Helper {
 
 	public Dataset createDataSet(Hdf5HelperData hData, boolean extend) throws NullPointerException {
 		int datatypeClass = hData.h5Datatype.getDatatypeClass();
-		int datatypeSize = hData.h5Datatype.getDatatypeSize();
+		int datatypeSize = (int) hData.h5Datatype.getDatatypeSize();
 		int dtype = HDF5Utils.getDtype(datatypeClass, datatypeSize);
 		int dims[] = new int[hData.dims.length];
 		for (int i = 0; i < hData.dims.length; i++)
